package refactors and reupdate trunk
