+++ src/main/java/net/coderbot/iris/shaderpack/ShaderPack.java
@@ -9,6 +9,8 @@ import java.util.Optional;
 import java.util.Properties;
 
 import net.coderbot.iris.Iris;
+import net.coderbot.iris.shaderpack.config.ShaderPackConfig;
+import net.coderbot.iris.shaderpack.parse.IdMap;
 import org.apache.logging.log4j.Level;
 import org.jetbrains.annotations.Nullable;
